[IMP] techwk3: Inherit existing report
Inheriting an existing report is more resilient to upgrades and
updates. It is also easier to parse, as only the most important
parts are included in the file itself.

A better version of this file might replace the existing t-if
attribute with different logic, but I'm pretty sure that the xpath
element can only add attributes, not edit or remove them.

opw-4986416