[IMP] techwk3: Adding product and partner fields
This code has been committed before testing, since I want a history
to return to if the data becomes corrupted.

Since the contact form must have an O-2-M field, that implies the
product record must also have a M-2-O field. So I added this to the
product record as well.

opw-4986419