[IMP] techwk3: Prune product search results.
This commit adds a method that removes searched products after they
are returned. This was added before I realized record rules exist
for products on the website_sales page.

I expect that this is slower than the record rule option, since I
assume that record rules are factored in with the main search
querry. I will do some testing.

This is the second requirement. I will complete the third using
record rules.

opw-4986419