[IMP] techk3: Redirect public users from shop to login
As stated, public users are redirected to the login page instead of
the store page. I believe there's a way to redirect the customer to
the shop after they log in... but this is enough for now.

The first requirement (see any product related to anyone in your
company) is still not met. Also, after speaking with Dirk and Pat,
I notice that products are still visible. The next commit should
improve this requirement.

opw-4986419