[IMP] techwk3: Add red text, fix logic
This commit should be fully functional, for the purposes of the
testing criteria. I only picked one sale order, as mentioned in
last commit, because the task only seems to care about the previous
sales order.

I also added logic to account for multiple lines of a product
with different price_units. Idk if this is common, but it's
possible.

opw-4986424