[IMP] techwk3: Implement barcode generation
I thought this would take more time, but apparently not!

There was an issue with the self.id and self.id.origin fields, but
I resolved that with some conditional logic. I was also surprised
to learn that the code works fine just by removing the readonly
restriction -- I thought it would forcibly recalculate.

I also note that the selection fields were defined to make the
compute method simpler. This might impact the readability of the
code though...

opw-4986423