[IMP] techwk3: Display results correctly
This commit adds the correct records to display the customer ref
and invoice_user_id fields. It was also necessary to edit the
results so that the data was formatted correctly.

I got the salesperson's name poorly, getting the invoice_user_id
within the SQL querry and then using a separate .browse command.
Alternatively, I could have added the following column:

	ARRAY_AGG(users.name) AS salesperson,

and the following left join:

	JOIN res_users users on user.id = move.invoice_user_id

then changed the result reference to salesperson. I did not do this
because this task no longer matters, and I should spend my time on
others.

opw-4986425