[ADD] techwk3: Update vendor pricelist to latest product price.
This was a pretty simple one, only one commit.

I split the vendor_pricelist search in two to ensure that only one
vendor pricelist is returned. This should prevent vendor pricelists
for a specific variant from being overwritten.

However, this code does not anticipate the possibility of multiple
vendor pricelists for the same vendor and product.

opw-4986426